pyproject.toml
[tool.poetry]
name = "accounts"
version = "0.1.0"
description = ""
authors = ["Ralph R. <10516663+rr0001@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyPDF2 = "^1.26.0"
openpyxl = "^3.0.5"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
